cabinos/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
cryptography
pynacl
python-jose
redis
